backend/requirements.txt
# Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
gunicorn==21.2.0

# Data Processing and ML
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1

# API and HTTP
requests==2.31.0
urllib3==2.0.4

# Environment and Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Database
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7

# Authentication
Flask-Login==0.6.2
Authlib==1.2.1

# Caching and Performance
redis==4.6.0
Flask-Caching==2.0.2

# Utilities
python-dateutil==2.8.2
Pillow==10.0.0
tqdm==4.66.1

# Development and Testing
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
